fix: extract bootstrapper and comment apps
