+++ .gitignore
@@ -10,7 +10,6 @@ out/
 .cache
 **/public/bootstrap/*.js.map
 **/public/bootstrap/*.js
-**/public/sw.js*
 cypress/screenshots/
 cypress/videos/
