+++ apps/boostrappers/index.html
@@ -9,7 +9,7 @@
   <body>
     <div>
       <div data-chirpy-comment="true"></div>
-      <script type="module" async src="/index.ts" data-chirpy-domain="chirpy.dev"></script>
+      <script type="module" async src="./src/index.ts" data-chirpy-domain="chirpy.dev"></script>
     </div>
   </body>
 </html>
