+++ apps/boostrappers/src/__tests__/comment.test.tsx
@@ -1,6 +1,6 @@
 import { cleanup, render } from '@testing-library/react';
 
-import { comment } from '../bootstrapper/comment/comment-bootstrapper';
+import { comment } from '../comment-bootstrapper';
 
 describe('comment', () => {
   beforeEach(() => {
