+++ apps/main/package.json
@@ -3,47 +3,19 @@
   "private": true,
   "scripts": {
     "dev": "next dev",
-    "dev:both": "concurrently 'next dev' 'npm run dev:comment'",
-    "dev:comment": "vite --config scripts/vite/comment.config.js",
-    "dev:sw": "VITE_DEBUG=true npm run build:sw",
     "debug": "NODE_OPTIONS='--inspect' next dev",
-    "build": "concurrently 'next build' 'npm run build:comment' 'npm run build:sw'",
-    "build:comment": "vite build --config scripts/vite/comment.config.js",
-    "build:sw": "vite build --config scripts/vite/sw.config.js",
-    "build:comment:dev": "VITE_DEBUG=true npm run build:comment",
-    "build:sw:dev": "VITE_DEBUG=true npm run build:sw",
-    "build:local": "dotenv -e .env.local -- concurrently 'next build' 'npm run build:comment:dev' 'npm run build:sw:dev'",
+    "build": "next build",
+    "build:local": "dotenv -e .env.local -- next build",
     "analyze": "ANALYZE=true npm run build:local",
     "start": "next start",
     "start:local": "dotenv -e .env.local -- next start",
     "postinstall": "patch-package",
     "g:gql": "graphql-codegen --config ./scripts/graphql-codegen/server-codegen.yml && graphql-codegen --config ./scripts/graphql-codegen/client-codegen.yml",
     "g:gql:watch": "concurrently 'graphql-codegen --watch --config ./scripts/graphql-codegen/server-codegen.yml' 'graphql-codegen --config ./scripts/graphql-codegen/client-codegen.yml --watch'",
-    "cy:open": "CYPRESS_BASE_URL=http://localhost:3000 cypress open",
-    "cy:run": "CYPRESS_BASE_URL=http://localhost:3000 cypress run",
     "ut": "DEBUG_PRINT_LIMIT=10000 jest --silent=false",
     "ut:coverage": "DEBUG_PRINT_LIMIT=10000 jest --silent=false --coverage",
-    "lint": "eslint ./",
     "sb": "start-storybook -p 6006 -s ./public",
     "build-sb": "build-storybook -s ./public"
-  },
-  "simple-git-hooks": {
-    "pre-commit": "npx lint-staged"
-  },
-  "lint-staged": {
-    "*.{json,md,scss,css}": [
-      "prettier --write"
-    ],
-    "**/generated/**/*.{ts,tsx}": [
-      "prettier --write"
-    ],
-    "*.graphql": [
-      "prettier --write",
-      "npm run g:gql"
-    ],
-    "*.{ts,tsx,js}": [
-      "eslint --fix"
-    ]
   },
   "dependencies": {
     "@geist-ui/react-icons": "1.0.1",
@@ -150,13 +122,6 @@
     "dotenv": "16.0.0",
     "dotenv-cli": "5.1.0",
     "dotenv-expand": "8.0.3",
-    "eslint": "8.14.0",
-    "eslint-config-next": "12.1.5",
-    "eslint-config-prettier": "8.5.0",
-    "eslint-plugin-jest": "26.1.5",
-    "eslint-plugin-prettier": "4.0.0",
-    "eslint-plugin-storybook": "0.5.11",
-    "eslint-plugin-unicorn": "42.0.0",
     "intersection-observer": "0.12.0",
     "jest": "28.0.3",
     "jest-environment-jsdom": "28.0.2",
@@ -172,53 +137,7 @@
     "storybook-addon-next": "1.6.2",
     "tsconfig-paths-webpack-plugin": "3.5.2",
     "typescript": "4.6.4",
-    "turbo": "1.2.8",
-    "vite": "2.9.6",
     "webpack": "5.72.0",
     "whatwg-fetch": "3.6.2"
-  },
-  "engines": {
-    "node": "14.x"
-  },
-  "renovate": {
-    "extends": [
-      "config:base",
-      "docker:disable"
-    ],
-    "packageRules": [
-      {
-        "matchPackagePatterns": "^@tiptap/",
-        "groupName": [
-          "Tiptap packages"
-        ]
-      },
-      {
-        "matchPackagePatterns": "^dotenv",
-        "groupName": [
-          "Dotenv packages"
-        ]
-      },
-      {
-        "matchDepTypes": [
-          "devDependencies"
-        ],
-        "prPriority": -1
-      },
-      {
-        "matchPackagePatterns": [
-          "^react",
-          "^next"
-        ],
-        "prPriority": 5
-      }
-    ],
-    "labels": [
-      "📦 package"
-    ]
-  },
-  "nextBundleAnalysis": {
-    "budget": 256000,
-    "budgetPercentIncreaseRed": 10,
-    "showDetails": true
   }
 }
