+++ apps/main/public/sw.js
@@ -1,0 +1,1 @@
+(function(a){typeof define=="function"&&define.amd?define(a):a()})(function(){"use strict";const a=self;function c(t){const i=new URL(t,a.location.origin).href;return a.clients.matchAll({type:"window",includeUncontrolled:!0}).then(e=>{let s=null;for(const r of e)if(r.url===i){s=r;break}return s?s.focus():a.clients.openWindow(i)})}const n=self;n.addEventListener("push",t=>{var e;if(((e=n.Notification)==null?void 0:e.permission)!=="granted"||!t.data)return;const i=t.data.json();t.waitUntil(n.registration.showNotification(i.title,{requireInteraction:!0,icon:"/favicon.png",data:i,body:`"${i.body}"`,tag:i.url}))}),n.addEventListener("notificationclick",t=>{const i=t.notification,e=t.notification.data;i.close();const s=c(e.url);t.waitUntil(s)});const o="1651994312269";n.addEventListener("install",t=>{t.waitUntil(caches.open(o).then(i=>i.addAll(["/"])))}),n.addEventListener("fetch",t=>{t.respondWith((async()=>t.request.mode==="navigate"&&t.request.method==="GET"&&n.registration.waiting&&(await n.clients.matchAll()).length<2?(n.registration.waiting.postMessage("skipWaiting"),new Response("",{headers:{Refresh:"0"}})):await caches.match(t.request)||fetch(t.request))())}),n.addEventListener("activate",t=>{t.waitUntil(caches.keys().then(i=>Promise.all(i.map(e=>{if(e!=o)return caches.delete(e)}))))}),n.addEventListener("message",t=>{t.data==="skipWaiting"&&n.skipWaiting()})});
