+++ apps/service-worker/index.html
@@ -1,0 +1,14 @@
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Demo</title>
+  </head>
+
+  <body>
+    <div>
+      <script type="module" async src="./src/index.ts" data-chirpy-domain="chirpy.dev"></script>
+    </div>
+  </body>
+</html>
