+++ apps/service-worker/package.json
@@ -1,0 +1,19 @@
+{
+  "name": "@chirpy/service-worker",
+  "version": "0.0.0",
+  "private": true,
+  "scripts": {
+    "build": "vite build",
+    "build:local": "dotenv -e ../main/.env.local vite build"
+  },
+  "license": "MIT",
+  "devDependencies": {
+    "dotenv": "16.0.0",
+    "dotenv-cli": "5.1.0",
+    "dotenv-expand": "8.0.3",
+    "vite": "2.9.6"
+  },
+  "publishConfig": {
+    "access": "public"
+  }
+}
