+++ package.json
@@ -2,29 +2,9 @@
   "name": "chirpy",
   "private": true,
   "scripts": {
-    "dev": "next dev",
-    "dev:both": "concurrently 'next dev' 'npm run dev:comment'",
-    "dev:comment": "vite --config scripts/vite/comment.config.js",
-    "dev:sw": "VITE_DEBUG=true npm run build:sw",
-    "debug": "NODE_OPTIONS='--inspect' next dev",
-    "build": "concurrently 'next build' 'npm run build:comment' 'npm run build:sw'",
-    "build:comment": "vite build --config scripts/vite/comment.config.js",
-    "build:sw": "vite build --config scripts/vite/sw.config.js",
-    "build:comment:dev": "VITE_DEBUG=true npm run build:comment",
-    "build:sw:dev": "VITE_DEBUG=true npm run build:sw",
-    "build:local": "dotenv -e .env.local -- concurrently 'next build' 'npm run build:comment:dev' 'npm run build:sw:dev'",
-    "analyze": "ANALYZE=true npm run build:local",
-    "start": "next start",
-    "start:local": "dotenv -e .env.local -- next start",
-    "postinstall": "patch-package",
-    "g:gql": "graphql-codegen --config ./scripts/graphql-codegen/server-codegen.yml && graphql-codegen --config ./scripts/graphql-codegen/client-codegen.yml",
-    "g:gql:watch": "concurrently 'graphql-codegen --watch --config ./scripts/graphql-codegen/server-codegen.yml' 'graphql-codegen --config ./scripts/graphql-codegen/client-codegen.yml --watch'",
     "cy:open": "CYPRESS_BASE_URL=http://localhost:3000 cypress open",
     "cy:run": "CYPRESS_BASE_URL=http://localhost:3000 cypress run",
-    "ut": "DEBUG_PRINT_LIMIT=10000 jest --silent=false",
-    "ut:coverage": "DEBUG_PRINT_LIMIT=10000 jest --silent=false --coverage",
     "lint": "eslint ./",
-    "sb": "start-storybook -p 6006 -s ./public",
     "build-sb": "build-storybook -s ./public"
   },
   "simple-git-hooks": {
@@ -44,65 +24,6 @@
     "*.{ts,tsx,js}": [
       "eslint --fix"
     ]
-  },
-  "dependencies": {
-    "@geist-ui/react-icons": "1.0.1",
-    "@headlessui/react": "1.6.0",
-    "@next/bundle-analyzer": "12.1.5",
-    "@radix-ui/colors": "0.1.8",
-    "@react-aria/button": "3.4.3",
-    "@sendinblue/client": "3.1.0",
-    "@tailwindcss/typography": "0.4.1",
-    "@testing-library/cypress": "8.0.2",
-    "@tiptap/extension-link": "2.0.0-beta.37",
-    "@tiptap/extension-placeholder": "2.0.0-beta.48",
-    "@tiptap/extension-underline": "2.0.0-beta.23",
-    "@tiptap/react": "2.0.0-beta.109",
-    "@tiptap/starter-kit": "2.0.0-beta.184",
-    "canvas-confetti": "1.5.1",
-    "chart.js": "3.7.1",
-    "clsx": "1.1.1",
-    "cors": "2.8.5",
-    "csstype": "3.0.11",
-    "d3": "3.5.17",
-    "datamaps": "0.5.9",
-    "dayjs": "1.11.1",
-    "debounce-promise": "3.1.2",
-    "eta": "1.12.3",
-    "framer-motion": "6.3.3",
-    "github-slugger": "1.4.0",
-    "graphql": "16.4.0",
-    "graphql-ws": "5.8.1",
-    "gray-matter": "4.0.3",
-    "html-loader": "3.1.0",
-    "jsonwebtoken": "8.5.1",
-    "lodash": "4.17.21",
-    "mdx-prism": "0.3.4",
-    "next": "12.1.5",
-    "next-auth": "4.3.4",
-    "next-compose-plugins": "2.2.1",
-    "next-connect": "0.12.2",
-    "next-mdx-remote": "3.0.8",
-    "next-plausible": "3.1.9",
-    "next-themes": "0.1.1",
-    "next-urql": "3.3.2",
-    "nodemailer": "6.7.4",
-    "patch-package": "6.4.7",
-    "react": "18.1.0",
-    "react-dom": "18.1.0",
-    "react-error-boundary": "3.1.4",
-    "react-flatpickr": "3.10.11",
-    "react-flip-move": "3.0.4",
-    "reading-time": "1.5.0",
-    "rehype": "12.0.1",
-    "super-tiny-icons": "0.4.0",
-    "superjson": "1.9.0",
-    "tailwindcss": "3.0.24",
-    "type-fest": "2.12.2",
-    "urql": "2.2.0",
-    "web-push": "3.4.5",
-    "wonka": "4.0.15",
-    "ws": "8.6.0"
   },
   "devDependencies": {
     "@graphql-codegen/cli": "2.6.2",
